+++ src/SortJSON/CodeActions/FixAllProvider.ts
@@ -1,11 +1,13 @@
 import * as vscode from "vscode";
+import { Settings } from '../../Settings';
+import { compareFileName } from '../../utils';
 
 export class SortJSONFixAllProvider implements vscode.CodeActionProvider {
     static fixAllCodeActionKind = vscode.CodeActionKind.SourceFixAll.append("sort-json");
 
     static metadata = { providedCodeActionKinds: [SortJSONFixAllProvider.fixAllCodeActionKind] };
 
-    async provideCodeActions(document, _range, context, _token) {
+    async provideCodeActions(_document, _range, context, _token) {
         if (!context.only) return [];
 
         if (
@@ -14,6 +16,12 @@ export class SortJSONFixAllProvider implements vscode.CodeActionProvider {
         ) {
             return [];
         }
+
+        const filesToIgnore = Settings.ignoreFiles;
+        const shouldIgnoreFix = filesToIgnore.some(compareFileName);
+
+        if (shouldIgnoreFix) return [];
+
 
         const isFixAllAction = true; // If false it shows the collection key prompt and info notification.
         const fixAllAction = await vscode.commands.executeCommand("sort-json.ascendingSort", undefined, isFixAllAction);